+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Changelog
   * gpxdo: dates may now contain a time
   * gpxdo split --stops MINUTES instead of fix --jumps
   * gpxdo fix --clear-times, --set-first-time, --set-last-time, --add-times
+  * new: Track.locate_point() and Track.add_locations()
 
 1.6.0 release 2019-01-xx
 ------------------------